clean-schema-registry/requirements.txt
pyyaml
requests
tabulate
